Switch `FeedStore` to its own file
